[#3] refactor: 사용 dto 변경
