--- src/main/java/com/napzak/global/common/dto/BaseResponse.java
@@ -1,8 +1,0 @@
-package com.napzak.global.common.dto;
-
-import org.springframework.http.HttpStatus;
-
-public interface BaseResponse {
-    HttpStatus status();
-    String message();
-}
